Disambiguate env vars for model download from Github-reserved values
When running in a Github Action runner, `GITHUB_REPOSITORY` is a reserved environment variable that ideally should not be overridden for fear of erratic/unexpected side-effects.

This commit renames the environment variables to be a bit more distinct and related to their purpose within spleeter i.e. for building a model download destination.

Fixes https://github.com/deezer/spleeter/issues/781